+++ scripts/run_langgraph_nodes/data/results/analyze_code_chunk_result.json
@@ -1,0 +1,11 @@
+{
+  "chunk_id": "data_gathering_and_processing",
+  "summary": "This code implements a data gathering and processing system for various agricultural and environmental grants. It fetches grant information from multiple sources (farming finder, woodland grants, vet visits, woodland offer), processes the data, generates embeddings and summaries using OpenAI services, and uploads the processed information to Azure AI Search. The system also manages manifests in S3 to track processed grants and handle updates.",
+  "data_model": "The main data entities in this system are:\n\n1. Grant\n   - title: string\n   - content: string\n   - updateDate: Date\n   - url: string\n\n2. Manifest\n   - link: string\n   - lastModified: string\n   - documentKeys: string[]\n   - summariesKeys: string[]\n\n3. DocumentChunk\n   - chunk_id: string\n   - parent_id: string\n   - chunk: string\n   - title: string\n   - grant_scheme_name: string\n   - source_url: string\n   - content_vector: number[]\n\nData flow:\n1. Grants are fetched from various sources\n2. Grants are processed and chunked\n3. Embeddings and summaries are generated for each chunk\n4. Processed data is uploaded to Azure AI Search\n5. Manifests are updated and stored in S3\n\n```mermaid\nerDiagram\n    Grant ||--|{ DocumentChunk : \"chunked into\"\n    Grant ||--|| Manifest : \"tracked by\"\n    DocumentChunk ||--|| AzureAISearch : \"uploaded to\"\n    Manifest }|--|| S3 : \"stored in\"\n```",
+  "interfaces": "1. API Endpoints:\n   - POST /gather-data: Triggers the data gathering and processing workflow\n\n2. External Service Interfaces:\n   - GOV.UK API: Fetches grant content\n   - Azure AI Search: Uploads processed document chunks and summaries\n   - OpenAI API: Generates embeddings and summaries\n   - AWS S3: Stores and retrieves manifests\n\n3. Internal Service Interfaces:\n   - FarmingFinder: Fetches grants from \"Find funding for land or farms\" finder\n   - WoodlandGrants: Fetches woodland grant information\n   - VetVisits: Fetches vet visit grant information\n   - WoodlandOffer: Fetches woodland offer grant information",
+  "business_logic": "1. Grant Processing Workflow:\n   - Fetch grants from various sources (farming finder, woodland grants, vet visits, woodland offer)\n   - Check if grants are new or updated compared to the manifest\n   - Process new or updated grants:\n     - Chunk grant content\n     - Generate embeddings for each chunk using OpenAI\n     - Generate short summaries for grants using OpenAI\n   - Upload processed chunks and summaries to Azure AI Search\n   - Update manifest with processed grant information\n   - Remove outdated grants from Azure AI Search\n\n2. Chunking Algorithm:\n   - Split long documents into smaller chunks based on token limit\n   - Ensure chunks end at sentence boundaries when possible\n   - Add metadata (title, grant scheme name, source URL, chunk number) to each chunk\n\n3. Manifest Management:\n   - Track processed grants with their last modified dates\n   - Use manifests to determine which grants need processing\n   - Store manifests in S3 for persistence across runs",
+  "dependencies": "1. External Libraries:\n   - @azure/search-documents: Interacting with Azure AI Search\n   - @langchain/openai: OpenAI API wrapper for embeddings and chat\n   - @aws-sdk/client-s3: AWS S3 client for manifest storage\n   - js-tiktoken: Token counting for text chunking\n   - turndown: HTML to Markdown conversion\n   - cheerio: HTML parsing for woodland grants\n\n2. External Services:\n   - Azure AI Search: Document storage and retrieval\n   - OpenAI API: Generating embeddings and summaries\n   - AWS S3: Storing and retrieving manifests\n   - GOV.UK API: Fetching grant content\n\n3. Internal Dependencies:\n   - config: Application configuration management\n   - logger: Logging utility\n   - proxy-fetch: HTTP requests with proxy support",
+  "configuration": "The system uses a configuration module (config) to manage various settings:\n\n1. Azure OpenAI Configuration:\n   - searchUrl: Azure AI Search URL\n   - indexName: Main index name\n   - summaryIndexName: Summary index name\n   - searchApiKey: API key for Azure AI Search\n   - openAiInstanceName: Azure OpenAI instance name\n   - openAiKey: Azure OpenAI API key\n\n2. AWS Configuration:\n   - region: AWS region\n   - s3bucket: S3 bucket name for storing manifests\n   - s3Endpoint: Optional custom S3 endpoint\n\n3. Grant Source URLs:\n   - farmingFinder.searchUrl: URL for farming finder grants\n   - farmingFinder.findFarmingUrl: Base URL for farming grants\n   - vetVisits.url: URL for vet visits grant\n   - woodlandCreation.url: URL for woodland creation grants\n   - woodlandOffer.url: URL for woodland offer grants\n\n4. Proxy Configuration:\n   - httpsProxy or httpProxy: Proxy URL for HTTP(S) requests\n\n5. Manifest File Names:\n   - Scheme-specific manifest file names for different grant types\n\nEnvironment variables are likely used to populate these configuration values securely.",
+  "infrastructure": null,
+  "non_functional": "1. Error Handling and Logging:\n   - Extensive use of try-catch blocks to handle errors\n   - Detailed error logging using a custom logger\n   - Graceful error recovery in some cases (e.g., fallback summary generation)\n\n2. Performance Considerations:\n   - Chunking of large documents to manage token limits\n   - Use of Azure AI Search for efficient storage and retrieval of processed data\n   - Parallel processing of multiple grant sources\n\n3. Security:\n   - Use of API keys for authentication with external services\n   - Proxy support for network requests\n\n4. Reliability:\n   - Manifest-based tracking to resume processing in case of interruptions\n   - Removal of outdated grants to maintain data accuracy\n\n5. Monitoring:\n   - Logging of important events and errors for monitoring purposes\n\n6. Data Privacy:\n   - Stripping of links from processed content to maintain privacy"
+}
