+++ scripts/run_langgraph_nodes/data/results/data_model_report_result.json
@@ -1,0 +1,11 @@
+{
+  "report_sections": {
+    "data_model": "# Data Model Report\n\n# Data Model Analysis Report\n\n## Logical Data Models and Entities\n\nThe codebase primarily deals with three main entities:\n\n1. Grant\n2. Manifest\n3. DocumentChunk\n\nAdditionally, there are two database collections:\n\n4. mongo-locks\n5. example-data\n\n## ERD Diagram\n\n```mermaid\nerDiagram\n    Grant ||--|{ DocumentChunk : \"chunked into\"\n    Grant ||--|| Manifest : \"tracked by\"\n    DocumentChunk }|--|| AzureAISearch : \"stored in\"\n    Manifest }|--|| S3 : \"stored in\"\n    MONGO-LOCKS {\n        string id\n    }\n    EXAMPLE-DATA {\n        string id\n    }\n```\n\n## Detailed Breakdown of Models\n\n### 1. Grant\n- Fields:\n  - title: string\n  - content: string\n  - updateDate: Date\n  - url: string\n\nRelationships:\n- One Grant can be chunked into many DocumentChunks\n- One Grant is tracked by one Manifest\n\n### 2. Manifest\n- Fields:\n  - link: string\n  - lastModified: string\n  - documentKeys: string[]\n  - summariesKeys: string[]\n\nRelationships:\n- One Manifest tracks one Grant\n- Manifests are stored in S3\n\n### 3. DocumentChunk\n- Fields:\n  - chunk_id: string\n  - parent_id: string\n  - chunk: string\n  - title: string\n  - grant_scheme_name: string\n  - source_url: string\n  - content_vector: number[]\n\nRelationships:\n- Many DocumentChunks are created from one Grant\n- DocumentChunks are stored in Azure AI Search\n\n### 4. mongo-locks (Collection)\n- Fields:\n  - id: string (indexed)\n\n### 5. example-data (Collection)\n- Fields:\n  - id: string (indexed)\n\n## Data Flow and Transformations\n\n1. Grants are fetched from various sources\n2. Grants are processed and chunked into DocumentChunks\n3. Embeddings and summaries are generated for each DocumentChunk\n4. Processed DocumentChunks are uploaded to Azure AI Search\n5. Manifests are updated to track the processed Grants and stored in S3\n\nThe data flow involves several transformations:\n- Grants are broken down into smaller DocumentChunks\n- Embeddings (content_vector) are generated for each DocumentChunk\n- Summaries are created for each DocumentChunk (not explicitly shown in the data model)\n\n## Data Validation and Integrity Checks\n\nThe provided code doesn't explicitly show data validation or integrity checks. However, some implicit checks can be inferred:\n\n1. The use of specific data types (e.g., string, Date) implies basic type checking.\n2. The creation of indexes on the 'id' field for both mongo-locks and example-data collections suggests that these fields are used for lookups and should be unique.\n3. The presence of 'updateDate' in the Grant model indicates that temporal consistency is tracked.\n\nTo improve data integrity, consider implementing the following:\n\n1. Validation of required fields for each model\n2. Checks for data format consistency (e.g., URL format for 'url' and 'source_url' fields)\n3. Ensuring referential integrity between related entities (e.g., DocumentChunk's parent_id should reference a valid Grant)\n4. Implementing version control or change tracking for Grants and Manifests\n5. Adding constraints to ensure the uniqueness of certain fields where appropriate (e.g., Grant's url, DocumentChunk's chunk_id)\n\nThe current data model provides a foundation for managing grants, their chunked representations, and tracking manifests. However, it would benefit from more explicit validation and integrity checks to ensure data consistency and reliability throughout the system.",
+    "interfaces": null,
+    "business_logic": null,
+    "dependencies": null,
+    "configuration": null,
+    "infrastructure": "",
+    "non_functional": null
+  }
+}
